Missed a few

+++ plugins/Xpressive/ExprSynth.cpp
@@ -398,7 +398,7 @@ struct sin_wave
 	static inline float process(float x)
 	{
 		x = positiveFraction(x);
-		return std::sinf(x * numbers::tau_v<float>);
+		return std::sin(x * numbers::tau_v<float>);
 	}
 };
 static freefunc1<float,sin_wave,true> sin_wave_func;
